backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pandas>=2.2.0
openpyxl>=3.1.0
python-multipart>=0.0.9
psycopg[binary]>=3.2.0
psycopg-pool>=3.2.0
python-dotenv>=1.0.1
msal>=1.31.0
httpx>=0.28.0
